feat: move tx history
