+++ packages/scripts/package.json
@@ -17,7 +17,7 @@
   "dependencies": {
     "@actions/core": "^1.10.1",
     "@actions/github": "^6.0.0",
-    "@arbitrum/sdk": "^4.0.1",
+    "@arbitrum/sdk": "^4.0.2",
     "@octokit/rest": "^21.0.2",
     "axios": "^1.7.7",
     "commander": "^12.1.0",
